added some exception + access control

+++ src/main/java/tn/fst/spring/projet_spring/services/logistics/LivreurServiceImpl.java
@@ -69,6 +69,9 @@ public class LivreurServiceImpl implements ILivreurService {
 
     @Override
     public void removeLivreur(Long id) {
+        if (!livreurRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Livreur not found with id: " + id);
+        }
         livreurRepository.deleteById(id);
     }
